Changes on top of version upgrade
The previous CL hasn't been modified from its original state.

This CL makes myriad changes that resolve bats test failures on the
library upgrade CL. These generally fall into three categories:

1) Exact text matching with the output of kubectl get ... -o jsonpath.
The output format has changed across versions. Fix: use jq -r  as that
output format is consistent.

2) Not waiting long enough for the repository to sync. Or, wait::for
instead of waiting for the repository to sync. Fix: increase wait::for
times for waiting for repos to sync, and remove wait::for on operations
which shouldn't need to wait once the repository reports synced.

3) Not properly cleaning up before tests. Some tests don't clean up
state they require as a precondition. Thus, they will fail continuously
if they ever exit unexpectedly. Fix: clean up this precondition state.

Remove nomos vet checks from CRD tests as they don't actually test what
is intended, and the functionality is fully covered in unit tests. If we
want an explicit E2E test for nomos vet's behavior with CRDs/CRs, we
don't need to use a cluster for it. As-before was passing as we were
only checking that nomos vet failed, and weren't actually pointing it at
a nomos repository - so it couldn't have passed anyway.

Additionally, upgrade the Kind version and image versions, and make the
default 1.17 as that is the default GKE cluster version.

Bug: 173730710

Change-Id: I5d31cfff4483b4027847cc791e335e35b5ba05ea