Increase QPS for API Server rate limiter.
The upgrade from 1.18 to 1.19 not only started respecting our context
timeouts, but also prevents us from unintnetionally bypassing the rate
limiter. As it turns out for normal operation we far exceed the 5 QPS
defaults, so the rate limiter prevented us from being very active.

Mono-repo ran into this issue more than multi-repo as mono-repo has a
higher ratio of api calls per resource (since multi-repo is all
in-memory).

A QPS of 20 and burst of 40 seems to make things behave reasonably
again, so we've set that. It should be safe to increase this if
necessary.

Change-Id: Ieb1063dbc43d21ed64a7e520c10390be523f4052