Add a Reconciler service.
This service exposes port 8675 which is needed in order for Prometheus to scrape metrics from the Reconciler. The service config is added to the reconciler manager config map and mounted as a volume, similar to the deployment. Additional checks have been added to the unit/e2e tests to ensure that the service is created and destroyed properly.

Bug: 153128444

Change-Id: I78a4b40f65f7955d9f8afa1cff9dcc96cd869c83