Set up namespaced watches in namespace reconcilers
This PR also adds the support to allow a remediator to trigger a
parse-apply-watch loop when a watcher notices any objects which cannot
be managed by the reconciler. Without this support, if a root reconciler
takes over some resources managed by a namespace reconciler, the
namespace reconciler will not notice this until the next force-resync
(every hour) or the namespace repo is changed before the next
force-resync.

b/173435222

Change-Id: I6f50fb4ac95c1c81b2c329d77ba7f6b658f2c994