Re-sync the previously synced configs after reconciler restarts
Synced resources need re-sync after reconciler restarts because the new
reconciler might affect the sync status. One example is the impact of
the cluster-selector. For example, if the `CLUSTER_NAME` is changed in
the new reconciler, a different set of resources might be applied.
The current implementation won't re-sync imediately because both the
synced commit and the import commit are the same in a short period.

This commit adds an instantiation time for each related reconciler. It
compares the imported/synced time with the instantiation time and
triggers the re-sync operation.

The related reconcilers include:
1. pkg/importer/filesystem/reconcile.go: the importer translates the
imported files into ClusterConfig and NamespaceConfig resources and
updates them. Before, it only updates if .spec.resources is changed or
.spec.deleteSyncTime is set. Now, it also checks if the imported time
is NOT after the init time. If so, an update operation will be
performed.

2. pkg/syncer/reconcile/namespaceconfig.go,
pkg/syncer/reconcile/clusterconfig.go and pkg/syncer/crd/reconcile.go:
the ClusterConfig and NamespaceConfig reconcilers udpate the config
status if the resources are synced or failed to be synced. This commit
skips the update if the importTime is NOT after the init time.

3. pkg/syncer/reconcile/repostatus.go: the repostatus reconciler updates
the Repo status. Before, it considers a repo is synced only when the
sync token is the same as the import token, but the synced token could
be stale. This commit checks the sync time and sync state as well to
guarantee that the sync is up-to-date.

b/175227055

Change-Id: I02be69dd1fa4e85bb7070c27933e6cb2a3190819