Fix abs/rel path confusion in namespace tree
When adding the policy dir to the cmpath.Relative in ast.FileObjects, we
unintentionally started using the absolute path in the file system. This
broke running nomos vet on hierarchical repositories manually, but did
not impact ConfigSync.

This change makes the cmpath.Relative the original use - storing the
path relative to PolicyDir. PathAnnotationVisitor now requires PolicyDir
to be passed in at instantiation time so it can prefix these relative
paths with the PolicyDir.

TopLevelDirectoryValidator/ScopeTest no longer requires the PolicyDir,
so these are considerably shorter.

The core of this change is to pkg/importer/filesystem/reader.go, which
writes the path relative to PolicyDir into the FileObjects.

Bug: 173437956

Change-Id: Ifbc1d0a8c110ceb7ecc1a2b33e66f3b8eedff740