Convert v1 crds safely
Per go/crd-metadata, it isn't safe as-is to manually convert v1 CRDs to
v1beta1 CRDs. This CL modifies our importing logic to read v1 CRDs as
literal struct types instead of Unstructureds, then uses the generated
apiextensions functions to convert them to v1beta1 CRDs when we need to
get at CRD metadata.

This DOES NOT mean we are safe for 1.22. In 1.22 the v1beta1 definitions
(and conversion functions) will be removed, so before then we still need
to:
1) Use v1 CRDs in functions instead of v1beta1 CRDs.
2) Have some way of ensuring we remain compatible with v1beta1 CRDs.

See go/crd-metadata for full details.

This removes all unsafe conversions. So the CRDClusterConfig is now
parsed into literal v1beta1 CRDs (with safe conversion if declared as
v1). To avoid complex code duplication, this required splitting out file
reading/struct conversion logic into it's own package to avoid creating
a circular dependency.

Bug: 154527698

Change-Id: Iadc5ee74a3a947e25a8a76a719eca4ff79588c63