Return the same source error code for invalid git branches
If the git branch is invalid, the mono repo returns source error 2004,
but the rootsync and reposync return source error 2001. We should return
the same error code for both the mono repo and multi repos.

It also add more details on how to check the git-sync logs. Users have
to follow the instruction to check the logs in the git-sync container.
Otherwise, we need to persist and parse the git-sync logs.

b/137313920
b/143532335
b/135203862

Change-Id: Ie8d5b7476b10170060fd75be1291cf12dddb817d