Show pod logs on test failure
It can be difficult to debug problems when they occur in Prow but not
locally. Further, there are just a lot of pods to keep track of in
multi-repo so it isn't always obvious which one to check.

This CL makes the Go E2E tests print the logs of almost every pod on
failure, which will help with debugging.

This will be noisy if every test fails. It isn't clear what a good
workaround for this is. I've left out the fs-watcher logs as those can
get very noisy and rarely have useful information. Arguably git-sync
falls into this, but it'll help us see when we've misconfigured the git
repository so I've left those in.

Most of this CL is shifting around where constants are defined. There's
no logic changes to production code - only moving code. In some cases
I've exported constants that are frequently used. This required moving
many reconcilermanager/controller constants to
reconcilermanager/constants.go. In some cases this results in mono-repo
code depending on these constants, but that doesn't feel like a big deal
as we intend to delete mono-repo code once it's deprecated.

It would be cleaner to output these logs into ARTIFACTS, but I've tried
and failed to do so, and honestly it's better to have *something* than
to spend more time figuring out how to use ARTIFACTS.

Bug: 177572536

Change-Id: I597100fa7ed15da56256a1b725149fe10f51cfe1