Add Parser test for deprecated GK
Somehow we didn't already test this.

I was curious since the Admission Webhook may need to deal with
collisions across Group boundaries if we permit any set of versions of a
resource which have different Groups.

We'll need to either explicitly check for or forbid using ones that have
changed Group names. Obviously for non-core types we won't be able to do
the same thing.

Bug: 160787495
Change-Id: Ieb0cd45a8b4608198feb3de71d24709a559e36c3