Forbid extensions/v1beta1 Ingress
The networking v1 Ingress has been available since 1.14, and the
extensions one is no longer served as of 1.16. It is deprecated and will
be removed in 1.22, so we should encourage customers to upgrade.

Additionally, continuing to support extensions/v1beta1 Ingresses would
make the Webhook more difficult as users would be able to declare
Ingresses in a version which is not served.

Bug: 160787495
Change-Id: I2a43d15fc911ae669eabebdcacdbe45fb629c1e4