Implement declared field encoding and diffing.
There are two complementary parts to this CL. The first is in pkg/parse
which accepts an unstructured object and encodes all of its fields into
the smd (structured-merge-diff) format which server-side-apply uses.
This part will be integrated into the hydration flow for declared
resources so that we annotate them with their declared fields.

The second part is in pkg/webhook where we decode the declared fields
and use the smd libraries to generate a diff of the changed fields on an
object (passed from the admission request) and intersect those fields
with the decoded declared fields. The admission controller will reject a
non-empty intersection.

Bug: 160786928
Bug: 160786679
Change-Id: I065e3ee72aa83f34fbc15b0d3fbfe55ea092fdad