Spit readGitState into readGitCommitAndPolicyDir and readGitFiles
With this PR, the cache will cache the whole gitstate instead of only
caching policyDir.

`readGitCommitAndPolicyDir` will be called at the beginning of a
parse-apply-watch loop; `readGitFiles` will ONLY be called when the
gitstate in the cache is old.

This split is to further speed up the parse-apply-watch loop.

This PR also clean up the error handling in both `read` and `parse` so
that they both avoid logging any error encountered, instead return the
errors.

b/177554523

Change-Id: I2f9c831f96843c436012d4f866aaa3333aaab343