Add logic for merging Webhooks
Add MergeWebhookConfigurations, which merges the Webhook configuration
already on the API Server with the one required to match all of the
types declared in the repository.

Per go/nomos-acdc, we're not worrying about removing stale Rules/entries
as it introduces more complexity than we need now, and the cost of Rules
which are never matched is negligible.

Bug: 160787495

Change-Id: I1aa0e573d6da7d426276ebad816cd693b39569ab