parser: add owning-inventory for each resource
The kpt applier adds owning-inventory: <INVENTORY-ID> to each resource
being applied. We need to add the same owning-inventory annotation to
resources from parser. So that when remediator updates a resource,
the owning-inventory annotation is not removed.

This change also updates the applier used in remediator. When multi-repo
feature is enabled, the applier uses kubectl last-applied-configuration
annotation for the 3-way merge. When multi-repo feature is not enabled,
the applier uses the config sync last applied annotation.

Using kubectl last-applied-configuration annotation requires the
support in b/160786679. The admission webhook should block
users from changing the kubectl last-applied-configuration annotation
on managed resources.

Bug: b/177428720
Change-Id: I9e3e2bbf7d349d5ad236a0edcf79a14ebc5467b2