Make webhook logic apply to multi-repo
The previous CL only added logic for the webhook creation and updating
for mono-repo. This CL generalizes the logic for multi-repo.

This also uncovered the edge case where a CRD + CR are added in the same
CL, so I've also included logic that allows kindResourceMapper to add
types from both supported CRD versions.

After this there shouldn't be anything we can do here until we've got
the webhook implemented. Specifically, handling the edge cases where
multiple reconcilers try to modify the webhook simultaneously.

Bug: 160787495

Change-Id: Iba735026f96db400562053a3faa350b42d0b57a0