Stabilize the golang e2e test suite
This commit includes the following changes to stabilize the golang e2e
test suite:
- enable running tests in parallel only when test.parallel is greater
than 1. Otherwise, the test will fail when we switch the test contexts.
- increase the timeout from 100 minutes to 300 minutes as all test cases
share the same GKE cluster. We have to run them sequentially and make
sure we clean up the cluster before and after each test.
- add a flag to the cleanup function to prevent the after-test cleanup
function from failing on error as the before-test cleanup will check
again if the cluster is in a clean state.

b/169356141

Change-Id: I9df116fab188fe67e9f429e1fca1add7767c94c2