Replace the static Prometheus metric exporter with an OC Agent exporter.
The reconciler-manager service and reconciler service have been removed. Because pods are no longer exposing metrics on port 8675 and aren't exposing any network services for others to consume, they don't need a service.

The OpenTelemetry Agent has been added as a sidecar container to the reconciler-manager and reconcilers. It exports metrics to an OpenTelemetry Collector which is configured to export metrics to Prometheus and Stackdriver. The Collector config can be customized to send metrics to a different metric backend by applying a new `otel-collector-custom` ConfigMap and restarting the Collector deployments.

This also fixes the metrics unit tests that started failing after upgrading OpenCensus from v0.22.2 to v0.22.5.

Bug: 153128444
Bug: 176004486

Change-Id: I58d04c2fe0de99d5a7e2bf41c790ab62dcde93a5