Re-implement ObjectQueue as a rate-limited queue.
The initial implementation of the ObjectQueue tried to re-use as much of
the existing k8s workqueue logic as possible. This meant that we
essentially wrapped a rate-limited work queue of GVKNNs with an outer
ObjectQueue of objects mapped by GVKNN. This mostly worked except that
you have to have a yielding-lock mechanism so that callers can get
blocked on Get() while safely allowing concurrent Add() and Done()
callers to operate on the queue. The initial idea was to lock all of the
ObjectQueue *except* for the initial part of the Get() which called the
yielding-lock Get() of the underlying queue. I thought this was safe but
there was at least one race condition: b/178044278#comment8

I tried several ideas for how to keep the two queues in sync without
re-implementing the yielding lock on our ObjectQueue but I don't think
it is possible. So this CL implements the yielding lock on our queue and
makes all of the calls to it completely locked. The main catch is that
this required that our ObjectQueue implement the rate-limiting interface
so that it could properly signal to the yielding lock on a delayed
Add(). Rather than re-implementing the full interface, I added a
self-referential delaying interface within the ObjectQueue to itself.
This sounds more complicated than it is: the Retry() function will do a
rate-limited delay before calling the ObjectQueue's Add() function.

The use of the sync.Condition as a yielding lock is based on the
implementation of the underlying workqueue:
https://github.com/kubernetes/client-go/blob/master/util/workqueue/queue.go

Note that the underlying work queue is no longer rate-limited since the
outer ObjectQueue now handles that.

Bug: 178044278
Change-Id: I00b70a908ec64c412c21a7f88e419d31f7978c15