Make sure we update sync status when clearing sync errors.
If we encountered a sync error in update(), then it wouldn't get
updated/cleared on a subsequent run. I had a transient error which was
resolved on the next apply run. However because `syncStatusUpdated` was
still true, it skipped over the code which would have cleared out
`status.sync.errors`.

Bug: 179256891
Change-Id: Iae71c3fca980259d8e8d5c46fc84b6376a42c69e