Don't validate error message text in E2E
In general asking for string matching in tests leads to flakiness. It
should be a last resort when other options are infeasible.

Error messages change and get reworded regularly. Sometimes we end up
using different wording for multi/mono-repo, and we encourage developers
to update error messages regularly as behavior changes or to improve
the wording.

We don't want E2E tests to act as a change-detection tests for this as
we want E2E test failures to indicate functional problems with the
product.

It is possible that in some cases we want to differentiate between two
sub-errors in a tests. That is, to make sure that a particular version
of an error was returned, and not the other as the two have the same KNV
Error code. If this is the case and we NEED to test it in E2E, then we
should add additional error codes for those cases. The only time we
should consider it a test failure is when the error message tells the
user something wrong/incomplete, and we should verify that at the level
of error codes and not strings.

Change-Id: I314c30a5a9cfe2eaf5695c0166c8e7062763bca8