Fix use of crd_removal_validator to catch errors.
In mono repo, we were inspecting the wrong ClusterConfig to check for
declared CRDs. In multi repo, we were just ignoring declared CRDs
entirely. This was okay for namespace repo but a problem for root repo.
I added logic which they share. This will be a no-op for namespace repo
but also means we will be ready if/when we allow CRDs in namespace
repos.

Bug: 178219594
Change-Id: Ibdb1ccf5af7661d5f9ea5049a5939473880a9128