Avoid updating the `Status` field of RepoSync/RootSync unnecessarily
Currently, if the reconciler fails to reconcile a git change, it updates
the `Status` field of RepoSync/RootSync at the end of each
reconciliation, even if all these reconciliation returns the same
errors.

This PR adds a new data structure to track the `Status` field of
RepoSync/RootSync, and only sends requests to the APIServer to update
the `Status` field when the errors are different.

After this PR, there are two more improvements we can do:
1) avoid logging the same errors again and again;
2) introduce backoff retry. Currently, we retry every second.

b/179256891

Change-Id: I02419b67d98491b70048e87a0df02eedb141fa66