Make webhook valid
The code for generating and applying the webhook didn't quite work
right.

This CL fixes a lot of small issues like .SetName() not actually working
on non-pointer structs and giving each Webhook in the list a unique
name.

After all of these, I observe in tests that the Webhook does apply
successfully. This does not resolve the CRD ordering problem, which is
much more complex to solve.

Bug: 179216134

Change-Id: If381a2905dba0ff3857c193b1270b4faf6c1abb1