Add the test label to nomos-managed test namespaces
We are now running the golang e2e tests against persistent GKE clusters,
but some test namespaces are not cleaned up after the test, which causes
test failures in the next run. To be clear, those manually created
namespaces will be cleaned up because we use `namespace::create` which
adds the test label, but those nomos-managed namespaces that are defined
in the git repository don't have the test label and won't be cleaned up.
This commit adds the test label to the invocations of
`namespace::declare`, where we add the namespace directory and yaml to
the git repository. We are not adding the lable inside
`namespace::declare` because some test cases use the function to check
in non-testing namespaces, such as kube-system, which results in the
revert commit f00c633.

It also removes the "configmanagement.gke.io" prefix from the test
label because it is a preserved nomos prefix for nomos-managed
resources and using it will cause nomos sync error.

b/179075455

Change-Id: Ia8aee7b22cdae926184ab23b9859d4742e337cad