Exit e2e.sh with the exit code
The e2e.sh prints the test time in the end, which makes the script
always exit with 0 and makes the make target always succeed even when go
test fails.
This commit uses trap to print the test time and exit when error occurs.

This will avoid prow job showing green for failed tests.

b/169356141

Change-Id: Icb2016a7e2eb8a3d445979c739280c82fc8d3345