Refactor discovery Scoper to simplify usage.
This enables the new pkg/validate library to integrate much more cleanly
with the Scoper. I am still not satisfied with the errOnUnknown in the
Parser. I *think* that I can encapsulate that more cleanly, but I would
rather save that for a later CL since this one is already so large.

Bug: 178219594
Change-Id: Ic4e5c0d17d145c7ff04d56ac35a36a1c6942e380