Add Structured() method to FileObject.
With this method, we no longer need to convert certain objects into
their structured formats when we read them from Git. We can store all
objects in their unstructured format and generate the structured format
when necessary (eg to perform hydration and validation checks).

Note that one of the changes is to make Parse() return FileObjects
instead of Objects. This aligns with the work to separate parse from
validation since parse will need to pass FileObjects to validation and
it should be the one to return Objects.

Bug: 179532046
Change-Id: Ife534119e8e7b7ca2edc4cf626afdec6993d240c