Update libraries to k8s 1.20
DO NOT SUBMIT until 1.6.2 branch cut. This is pretty big and we'd rather
we let this simmer in case there are subtle bugs.

Also make code generation consistent across configsync.* and
configmanagement.* packages. Note that make generate now operates
on the package level rather than the file level, so we don't need all
the redundant object:generate=true.

This CL updates our internal k8s dependencies to 1.20. This is a better
time than later as since we will push this after a branch cut we'll have
longer to make sure everything is actually stable.

In 1.20 all of the k8s:deepcopy* directives are deprecated and must be
replaced with new ones.

No longer delete *generated* in "make clientgen" as that command doesn't
actually generate those files - that's what "make generate" does. Also
ensure we're actually getting the versions of the tools we say we're
getting by explicitly asking for the versions we want.

Documentation for code generation directives in
http://cl/356280305.

Also delete protobuf generation logic since we don't use it and
everything works fine without it.

Change-Id: I753071fad811a602dc90c35b05413062f1bba1a4