Introduce smart retry
Currently, if the reconciler fails to reconcile a git change, it will
retry every second, even if all these retries fail with the same errors.
The reconciler log will be filled with logs for the same errors.

This PR introduces smart retry:
 * For the first 5 retries for reconciling a git change which fail with
   the same errors, the reconciler waits 1 second before retrying;
 * For the remaining retries for reconciling a git change which fail
   with the same errors, the reconciler does exponential backoff retry
   up to 5 minutes (i.e., 1s, 2s, 4s, 8s, 16s, 32s, 64s, 128s, 256s, 5m, 5m, ...).

This change will greatly reduce the logging entries for the same errors.

b/179256891

Change-Id: Ie8e6ffee7d73027d6d4ae09da39ab7b8497dc168