Upgrade controller-runtime to v0.8.1
As with the other upgrade, this is an opportune moment to upgrade
libraries as we've just cut a release. This CL upgrades
controller-runtime to the most recent stable release.

The largest change here is from client.Client now requiring
client.Object instead of runtime.Object. This should simplify quite a
bit of our logic which requires runtime.Object/core.Object, but for now
I've just done a straight find+replace. In the future we won't have to
do those conversions, and we may want to eliminate core.Object since it
doesn't really provide anything now except a convenient
GroupVersionKind() method.

The new replace directive in go.mod is because Go really wants to
upgrade gnostic, but that requires gopkg.in/yaml.v3 which is
incompatible with kube-openapi.

The controller-runtime upgrade also highlights that we're using Events
in a way they aren't intended to be used. There may be a recommended way
to do what we want now - I've filed b/179816931 to track this but it
isn't important to change that now. What we have works.

This is all to prepare for the admission webhook. Rather than write it
and then upgrade the code, I'm upgrading now to avoid all of this later.

Bug: 179785113

Change-Id: I4d86db25387ecde01197e9483bf7d78bf5d348af