Swap the controllers' initialization order
This commit sets up the sync controller first, followed by the
git-importer controller. This is to fix an issue found in the golang e2e
test on the GKE cluster.

The git-importer reconciler updates the import time and import token and
sets the sync state to `stale`, while the cluster-config reconciler
updates the sync time and sync state (`synced` or `error`).

In some cases, if the git-importer reconciler updates the ClusterConfig
object before the cluster-config reconciler's init time, the
cluster-config reconciler will skip updating the sync status as it is
considered as a stalled config
(https://team.git.corp.google.com/gke-kubernetes-enterprise-control/prototype/+/refs/heads/master/pkg/syncer/reconcile/util.go#68).
Hence, the config will stay in the `stale` state and keep the commit in
the `in-progress` state forever.

If we swap the order, the import time is an obsolete time and the
cluster-config reconciler will skip updating the sync status. When the
git-importer reconciler loads the commit and updates the import time,
the import time will be after the cluster-reconciler's init time. The
cluster-config reconciler will update the sync time and sync status
correctly.

Why the cluster-config reconciler skips updating the previously imported
objects?
The ClusterConfig object might be created before the new git-importer
reconciler is up running (caused by a restart of the git-importer pod).
If the cluster-config reconciler updates the sync status to be `synced`,
any change in the ClusterConfig will not be applied. Therefore, the
cluster-config reconciler only updates the object imported after the
cluster-config reconciler is initialized.

Why we didn't see the issue in the test running againt the Kind cluster?
When running in the Kind cluster, the git-importer reconciler tries to
update the ClusterConfig object, but the /repo/root folder isn't
resolved yet by the git-sync container. When the /repo/root folder is
available, the import time is already after the cluster-config
reconciler's init time. However, while running in the GKE cluster, the
/repo/root folder is always resolved by the git-sync container.

b/179075455

Change-Id: I46c165416c9731124ba56288db74c3297770e2d8