Delete all clusters before running e2e tests
Otherwise we end up with dozens of clusters running simultaneously after
timeouts or other failures. It isn't clear if these clusters actually
get cleaned up with the Prow job.

We could say this is "best practice", but it doesn't look like it
has any impact on test runtime in clusters already struggling.

Change-Id: Ifba214c1e0e569c2748d436dd07de73b586ce823