Add e2e tests for multi-repo metrics.
This adds a new metrics testing library with functions that parse and validate the metrics recorded by the reconcilers. Documentation outlining the test framework and test patterns can be found in go/csmr-metrics-test-patterns.

Additional changes to the metrics:
1> The `watch_manager_updates` metric has been removed because the measurement value can be inferred from the `watch_manager_updates_duration_seconds` metric.
2> A new reconciler tag key has been added. The key name and value are dynamically sourced from the reconcilers pod name. This dynamic label is needed so that the OpenTelemetry Collector can differentiate between the different reconciler metrics. Otherwise metrics will be randomly sampled.

The new reconciler tag key is temporary until OpenTelemetry can support a configuration option in the OpenCensus exporter which allows resource attributes to be added to metric labels: https://github.com/open-telemetry/opentelemetry-collector/issues/2465. This would give the otel agents the ability to set the reconciler tag key before exporting the metrics to the collector.

Bug: 153128444
Bug: 157607685
Bug: 157588894
Bug: 162601559
Bug: 177923397

Change-Id: I1ef063be6db5d128f53080b526ffa04f17ad2b6b