Fix the intermittent unauthorized errors on the GKE cluster
The golang e2e tests running on the GKE cluster failed frequently due to
the following errors:
- error: You must be logged in to the server (the server has asked for
  the client to provide credentials)
- error: You must be logged in to the server (Unauthorized)

That is because the gcloud credentials expires every 60 minutes. The
credentials will refresh when necessary. However, we're copying the
~/.kube/config to a temp file. When it refreshes the creds, it updates
the token_expiry field in the temp file, not `~/.kube/config`. When the
next test case kicks in, it copies the unchanged `~/.kube/config` file
to a new temp file, which uses the expired credentials.

To avoid refreshing a temp file but copying a stalled kubeconfig, this
commit changes it to not dump the default kubeconfig to a temp file, but
set the KUBECONFIG environment variable when running on GKE clusters.
For the Kind clusters, it still generateds the KUBECONFIG file to a temp
directory, and we set the KUBECONFIG env variable as well.

b/179075455

Change-Id: Id41cce7c86622aa5466b49990a8d634ffc0e8f5f