Fix the jq error
We're seeing `jq: error (at <stdin>:27): Cannot iterate over null
(null)` errors when running the golang e2e test on the GKE clusters.
This is because repo.status.sync.inProgress is null, but we are still
iterating over the value. It's mysterious how the e2e tests get passed
against the Kind cluster, but we should make it work on the GKE cluster.

b/179075455

Change-Id: I8145aab641017d5c1db783bcc4fd67a47b56e4bf