Implement field diffing and validation in webhook.
This CL implements most of the validation checks for the webhook. It
checks that declared fields are not being modified and that Config Sync
metadata is not being modified. I also realized that we were only
handling resource updates, so I restructured the code to handle creates
and deletes as well.

Remaining work is to finish implementing Config Sync authz (including
manager precedence check).

Bug: 160786928
Bug: 160786679
Change-Id: I332391200952c0e263f9354b36cefc5ee2f71c11