Use namespace::create to create the gatekeeper-system namespace
basic.bats testcase 9 creates the gatekeeper-system namespace directly
using `kubectl create ns` and deletes the namespace before exiting the
test. If the test fails in the middle, the gatekeeper-system namespace
won't get a chance to be deleted.

This commit uses `namespace::create` to create the namespace with a
`testdata: true` label, so even the test fails in the middle, the
cleanup step will guarantee the testing namespace is deleted before the
next run.

b/179075455

Change-Id: I296162ff61f736ecfda7cabfedf2a55aac1abb2c