switch to kptapplier in reconciler
This change does the switch in reconciler to use the new kpt applier.
In the kpt applier, we switch to use the server side apply.

Alos update the client and base applier in remediator to use server
side apply. Now that both mono-repo and multi-repo modes switch to
server side apply.

As a side effect, the last-applied-config annotation and the
declared-config annotation are dropped.

Bug: b/177468879, b/175339528, b/174174388

Change-Id: Ie7126ab5b4ef1de48b482a9f0e16dabdd5fe0e1e