Force gcloud auth refresh before each test case
We're still seeing the intermittent unauthorized failures in the Prow
job after refreshing the access_token every 50 minutes. That is because
the refresh happened in the middle of a test case.

This commit forces gcloud to refresh the access_token before each test
case. Because we're using gcloud to refresh, it limits the remote
cluster to GKE cluster.

b/179075455

Change-Id: I072a4ab39deea6679959225be8f9e4d83ec671ad