Define new Raw and Scoped stages of validation.
This will be a slightly new approach to the new validation package. I
realized that the previous approach (with the FlatRoot and TreeRoot) had
some problems with when and how to introduce the Scoper. This new
approach does initial validation/hydration on the raw collection of
FileObjects and then builds a Scoper to generate a scoped collection of
FileObjects for the next round of validation/hydration.

For this first CL, I implemented the validation and hydration steps
around cluster selection from raw_parser.go. In the next CL I will move
the three commented out validation steps from the old FlatRoot/TreeRoot
approach to this new one. Then the third CL will introduce namespace
selection to the Scoped objects.

Bug: 178219594
Change-Id: I17a5d5375b0e0e81e6c294d80affd1238fb01ab4