Migrate duplicate name validation to final.
This CL introduces a "final" subpackage of the new validation package.
This subpackage is for validation that should occur after all other
validation and hydration is complete. Essentially it is for validating
the final state of the repo. In this specific case, we are validating
that no two objects have the same GKNN. We need to wait until after
hydration since objects get filtered and copied around which can affect
if there are duplicates or not.

Bug: 178219594
Change-Id: I1baf43f90372bca7ca1b4f01d5c05e7c786f88a8