Set git depth depending on sync revision for multirepo
A similar change for monorepo was merged in tg/951656. This commit
applies the same patch for multirepo.

It sets Git Depth to 1 if fetching from master to avoid
getting too many commits and OOMing the importer,
but gets 500 commits if a specific rev is mentioned
so that it is (hopefully) available in the history.

It may make sense to make rev fetching more intelligent/progressive
if 500 is not enough for some customers.

This addresses a customer issue at b/175088702.

Bug: 175334506
Change-Id: If93a4bbfc4648e1222adbb703b04a4a4a1f1855f