Migrate hierarchical namespace selector validation.
Note that there are two rounds of validation for hiearchical namespace
selectors. First we re-use the same validation done for unstructured
repos where we verify that selectors are only used for namespace-scoped
objects and not combined with metadata.namespace. Second, we do
validation specific for a hierarchical repo where we make sure that
legacy NamespaceSelectors only exist in abstract namespaces.

Bug: 178219594
Change-Id: I13e913d3603fe4915634ee4581e689de67e82024