Implement hierarhical namespace selector hydration
I realized that I made a mistake in the Tree builder: Namespaces and
NamespaceSelectors are cluster-scoped so they will be in that collection
of objects in the Scoped we build from. I fixed the Tree builder to
address this and made NamespaceSelectors an explicit field to assist
with validation and hydration involving them.

Bug: 178219594
Change-Id: I2e6a47d3ee7915b3d31f19ba94dec5169a1a2dfa