Switch legacy and multirepo to use pkg/validate
Note that this effectively eliminates the difference between Parser and
RawParser, so I removed the latter.

I also fixed an edge case in pkg/validate. When we perform cluster
selection, if a namespace gets filtered out then all of the objects in
that namespace get filtered as well, even if they don't specify a
cluster selector themselves. This is the same for hierarchical and
unstructured. However hieararchical repos allow objects to not specify
their namespace and instead "inherit" it from the namespace directory
they are located in. So I had to add a new ObjectNamespace hydrator that
sets the metadata namespaces for these objects *before* cluster
selection. This is only necessary for hierarchical repos; unstructured
repos only set a missing namespace on the "default namespace" which is
never cluster selected away.

Bug: 178219594
Change-Id: Ib3a98b800b7446d5c61f1e22fe5c9ae65fd27915