Add the ignore mutations lifecycle directive
This directive allows users to annotate objects in their repository to
indicate that they do not want us to maintain the state of the object in
their cluster once it exists.

We forbid users from modifying this annotation on managed objects.
There's a few edge cases I've noted in the bug, but they don't seem like
big issues. Generally - what is declared in the repository wins and
comes into effect once it is present on both the object and the
declaration.

Bug: 168726098

Change-Id: I7cb3a4c3385251617f532496832abf681b0b6f11